+++ docs/conf.py
@@ -112,15 +112,17 @@ pygments_style = "sphinx"
 #
 html_theme = "sphinx_rtd_theme"
 
+html_style = "css/customtheme.css"
+
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
 #
-# html_theme_options = {}
+html_theme_options = {"collapse_navigation": False}
 
 # The name of an image file (relative to this directory) to place at the top
 # of the sidebar.
-html_logo = "dbl.svg"
+html_logo = "topgg.svg"
 
 # The name of an image file (within the static path) to use as favicon of the
 # docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
@@ -130,7 +132,7 @@ html_favicon = "favicon.ico"
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-# html_static_path = ['_static']
+html_static_path = ["_static"]
 
 # Add any extra paths that contain custom files (such as robots.txt or
 # .htaccess) here, relative to this directory. These files are copied